Renamed constants and added softmax tab
