+++ src/detail_view/DetailImageTab.ts
@@ -1,11 +1,11 @@
-import {ADetailWindow} from './ADetailWindow';
+import {ADetailViewTab} from './ADetailViewTab';
 import {DataStoreCellSelection} from '../DataStore';
 
-export class DetailImageWindow extends ADetailWindow {
+export class DetailImageWindow extends ADetailViewTab {
   render() {
     // to implement
   }
-  
+
   clear() {
   }
 }
