+++ src/detail_view/SoftmaxStampTab.ts
@@ -1,0 +1,10 @@
+import {ADetailViewTab} from './ADetailViewTab';
+
+export class SoftmaxStampTab extends ADetailViewTab {
+  render() {
+
+  }
+
+  clear() {
+  }
+}
