--- src/detail_view/SoftmaxStampWindow.ts
@@ -1,10 +1,0 @@
-import {ADetailWindow} from './ADetailWindow';
-
-export class SoftmaxStampWindow extends ADetailWindow {
-  render() {
-
-  }
-  
-  clear() {
-  }
-}
